# api/openapi.yaml
openapi: 3.0.3
info:
  title: Notes Scheduler API
  description: API для управления заметками
  version: 1.0.0
servers:
  - url: http://localhost:8080
    description: Локальный сервер
paths:
  /api/v1/notes:
    post:
      summary: Создать заметку
      tags:
        - Notes
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateNoteRequest'
            example:
              title: "Купить продукты"
              tags: ["покупки", "срочно"]
      responses:
        '201':
          description: Заметка создана
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Note'
        '400':
          description: Ошибка валидации
    get:
      summary: Список активных заметок
      tags:
        - Notes
      responses:
        '200':
          description: Список заметок
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Note'
  /api/v1/notes/{noteId}/done:
    put:
      summary: Отметить заметку выполненной
      tags:
        - Notes
      parameters:
        - name: noteId
          in: path
          required: true
          schema:
            type: string
          example: "1704632400"
      responses:
        '200':
          description: Статус обновлён
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Note'
        '404':
          description: Заметка не найдена
components:
  schemas:
    CreateNoteRequest:
      type: object
      required:
        - title
      properties:
        title:
          type: string
          description: Заголовок заметки
          example: "Купить продукты"
        tags:
          type: array
          items:
            type: string
          example: ["покупки", "срочно"]
    Note:
      type: object
      properties:
        id:
          type: string
          description: Уникальный идентификатор
        title:
          type: string
        tags:
          type: array
          items:
            type: string
        status:
          type: string
          enum: [active, done]
          description: Статус заметки